move reactionroles actions into class file
