+++ src/index.ts
@@ -169,38 +169,13 @@ bot.on('guildMemberRemove', async member => {
     memberLeave(member)
 })
 
+import reactionRoles from './events/reactionRoles';
 bot.on('messageReactionAdd', async (reaction:Discord.MessageReaction, author:Discord.User) => {
-    let mongod = await MongoClient.connect(url, {'useUnifiedTopology': true});
-    let db = mongod.db(dbName);
-
-    let msg = await db.collection('msg').findOne({ _id: reaction.message.id })
-    if(!msg)return mongod.close();
-
-    let role = msg.roles.find((val:any) => val.emote == reaction.emoji.name)
-    if(!role)return mongod.close();
-
-    let member = reaction.message.guild.member(author)
-    if(!member)return mongod.close();
-    await member.roles.add(role.id)
-
-    return mongod.close();
+    reactionRoles.add(reaction, author);
 });
 
 bot.on('messageReactionRemove', async (reaction:Discord.MessageReaction, author:Discord.User) => {
-    let mongod = await MongoClient.connect(url, {'useUnifiedTopology': true});
-    let db = mongod.db(dbName);
-
-    let msg = await db.collection('msg').findOne({ _id: reaction.message.id })
-    if(!msg)return mongod.close();
-
-    let role = msg.roles.find((val:any) => val.emote == reaction.emoji.name)
-    if(!role)return mongod.close();
-
-    let member = reaction.message.guild.member(author)
-    if(!member)return mongod.close();
-    await member.roles.remove(role.id)
-
-    return mongod.close();
+    reactionRoles.remove(reaction, author);
 });
 
 bot.on('messageReactionAdd', async (reaction:Discord.MessageReaction, author:Discord.User) => {
